rename_module: Rename all packages in one shot avoid compilation errors
