+++ modules/windows/collect_ad.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package wmi
+package windows
 
 import "github.com/netdata/go.d.plugin/pkg/prometheus"
